server: treat the pty as finished if reading from it would block
previously we only consider pty finished if reading from it returned
error. this only happens if all processes on the other end of the pty
have exited. this means if the process we spawn spawned something that
does not exit when the process exits, job-security will hang waiting for
more data from the pty.

instead we should consider the pty as finished if reading from it would
block after the process we spawned exited. this still means if something
is flooding the pty with data after the main process exited, we
job-security client will keep reading. unsure what's the best thing to
do in that case.